fixup! fixup! feat(serve): add memory manager
