+++ docs/integrations/a2a.mdx
@@ -78,7 +78,7 @@ A2AServer lets you expose agents built in the BeeAI framework via A2A protocol.
 	from beeai_framework.agents.experimental import RequirementAgent
 	from beeai_framework.backend import ChatModel
 	from beeai_framework.memory import UnconstrainedMemory
-	from beeai_framework.serve.lru_memory_manager import LRUMemoryManager
+	from beeai_framework.serve.utils import LRUMemoryManager
 	from beeai_framework.tools.search.duckduckgo import DuckDuckGoSearchTool
 	from beeai_framework.tools.weather import OpenMeteoTool
 	
@@ -94,7 +94,8 @@ A2AServer lets you expose agents built in the BeeAI framework via A2A protocol.
 	    # Register the agent with the A2A server and run the HTTP server
 	    # For the ToolCallingAgent, we dont need to specify ACPAgent factory method
 	    # because it is already registered in the A2AServer
-	    A2AServer(config=A2AServerConfig(port=9999), memory_manager=LRUMemoryManager(maxsize=20)).register(agent).serve()
+	    # we use LRU memory manager to keep limited amount of sessions in the memory
+	    A2AServer(config=A2AServerConfig(port=9999), memory_manager=LRUMemoryManager(maxsize=100)).register(agent).serve()
 	
 	
 	if __name__ == "__main__":
