+++ docs/integrations/watsonx-orchestrate.mdx
@@ -91,6 +91,7 @@ This enables you to register and use your local BeeAI agents as external agents 
 	from beeai_framework.agents.experimental import RequirementAgent
 	from beeai_framework.backend import ChatModel
 	from beeai_framework.memory import UnconstrainedMemory
+	from beeai_framework.serve.utils import LRUMemoryManager
 	from beeai_framework.tools.weather import OpenMeteoTool
 	
 	
@@ -99,7 +100,8 @@ This enables you to register and use your local BeeAI agents as external agents 
 	    agent = RequirementAgent(llm=llm, tools=[OpenMeteoTool()], memory=UnconstrainedMemory(), role="a weather agent")
 	
 	    config = WatsonxOrchestrateServerConfig(port=8080, host="0.0.0.0", api_key=None)  # optional
-	    server = WatsonxOrchestrateServer(config=config)
+	    # use LRU memory manager to keep limited amount of sessions in the memory
+	    server = WatsonxOrchestrateServer(config=config, memory_manager=LRUMemoryManager(maxsize=100))
 	    server.register(agent)
 	
 	    # start an API with /chat/completions endpoint which is compatible with Watsonx Orchestrate
