+++ python/beeai_framework/serve/__init__.py
@@ -1,7 +1,7 @@
 # Copyright 2025 © BeeAI a Series of LF Projects, LLC
 # SPDX-License-Identifier: Apache-2.0
-from beeai_framework.serve.memory_manager import MemoryManager
+
 from beeai_framework.serve.server import Server
-from beeai_framework.serve.utils import init_agent_memory
+from beeai_framework.serve.utils import MemoryManager, init_agent_memory
 
 __all__ = ["MemoryManager", "Server", "init_agent_memory"]
