+++ python/beeai_framework/serve/server.py
@@ -9,8 +9,7 @@ from pydantic import BaseModel
 from typing_extensions import TypeVar
 
 from beeai_framework.serve.errors import FactoryAlreadyRegisteredError
-from beeai_framework.serve.lru_memory_manager import LRUMemoryManager
-from beeai_framework.serve.memory_manager import MemoryManager
+from beeai_framework.serve.utils import MemoryManager, UnlimitedMemoryManager
 
 TInput = TypeVar("TInput", bound=object, default=object, contravariant=True)
 TInternal = TypeVar("TInternal", bound=object, default=object)
@@ -24,7 +23,7 @@ class Server(Generic[TInput, TInternal, TConfig], ABC):
     def __init__(self, *, config: TConfig, memory_manager: MemoryManager | None) -> None:
         self._members: list[TInput] = []
         self._config = config
-        self._memory_manager: MemoryManager = memory_manager or LRUMemoryManager(maxsize=100)
+        self._memory_manager: MemoryManager = memory_manager or UnlimitedMemoryManager()
 
     @classmethod
     def register_factory(
