+++ python/beeai_framework/serve/utils.py
@@ -1,12 +1,51 @@
 # Copyright 2025 © BeeAI a Series of LF Projects, LLC
 # SPDX-License-Identifier: Apache-2.0
+from collections.abc import Callable
+from typing import Protocol
+
+from cachetools import LRUCache
 
 from beeai_framework.agents import AnyAgent
 from beeai_framework.logger import Logger
 from beeai_framework.memory import BaseMemory
-from beeai_framework.serve.memory_manager import MemoryManager
 
 logger = Logger(__name__)
+
+
+class MemoryManager(Protocol):
+    async def set(self, key: str, value: BaseMemory) -> None: ...
+
+    async def get(self, key: str) -> BaseMemory: ...
+
+    async def contains(self, key: str) -> bool: ...
+
+
+class UnlimitedMemoryManager(MemoryManager):
+    def __init__(self) -> None:
+        self._memory: dict[str, BaseMemory] = {}
+
+    async def set(self, key: str, value: BaseMemory) -> None:
+        self._memory[key] = value
+
+    async def get(self, key: str) -> BaseMemory:
+        return self._memory[key]
+
+    async def contains(self, key: str) -> bool:
+        return key in self._memory
+
+
+class LRUMemoryManager(MemoryManager):
+    def __init__(self, maxsize: int, getsizeof: Callable[[BaseMemory], int] | None = None) -> None:
+        self._cache: LRUCache[str, BaseMemory] = LRUCache(maxsize, getsizeof)
+
+    async def set(self, key: str, value: BaseMemory) -> None:
+        self._cache[key] = value
+
+    async def get(self, key: str) -> BaseMemory:
+        return self._cache[key]
+
+    async def contains(self, key: str) -> bool:
+        return key in self._cache
 
 
 async def init_agent_memory(
